Create many records in db with model.create()
Tp create many records in db use model.create() function. which takes array of objects with proper data as first argument and second argument is that function with done() method if there is no error, as per function of createManyPeople() function from server.js file.

Note :-
In database page of mongoose altas, the plural form of model name which we created is appears. So if model name is "Person" then "People" will appear in database collection of mongoose altas.