Use ModelName.findByIdAnRemove() or ModelName.findOneAndRemove()
These method are used to remove document matching conditio from database.
These methods have three arguments like:- (condition, [options], [callback])
In first arg, the condition is given to find that object.
The second arg is optional, it is use to set some optional object.
And the third is the callback function

Note:-
If you do not want to set options arg you can skipped it .